refactor: Move the ScsDataClient into momento, make it private.
We should not be publishing this internal object.
There's no reason for it to be in its own package, it makes it easier
to work with inside momento; the momento requests are intimately tied
to the data client.

Now the requests don't need the SimpleCacheClient wrapper at all.
